+++ test/example/php/return_types.php
@@ -1,0 +1,8 @@
+<?php
+function getSomeThing(): Thing {
+    return true ? something : somethingelse;
+}
+
+class Finder {
+    public function getSomeThing(): Thing {}
+}